Complete Backend and Frontend init - 9/14

+++ Backend/controller/product.controller.js
@@ -50,11 +50,14 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({ success: false, message: "ID not found" });
+  }
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ success: true, message: "Deleted" });
   } catch (error) {
     console.log(`Error. ${error.message}`);
-    res.status(404);
+    res.status(500).json({ success: false, message: "Error deleting Product" });
   }
 };
